==> Notebooks 2/5 - Parsing.ipynb <==
# **5 - Parsing**

## Import Libraries

import csv
import json

## Parse JSON


f = open('../Data 2/data_2.json', 'r')
j = json.load(f)
f.close()

sales = [] 
stores = []
customers = []
line_items = []

for sale in j['sales']:

    # Sales
    sale_dictionary = {}
    sale_dictionary['sale_id'] = sale['sale_id']
    sale_dictionary['date'] = sale['sale_date']
    sale_dictionary['sub_total'] = sale['sub_total']
    sale_dictionary['tax'] = sale['tax']
    sale_dictionary['total'] = sale['total_amount']
    sales.append(sale_dictionary)
    
    # Stores
    store_dictionary = {}
    store_dictionary['sale_id'] = sale['sale_id']
    store_dictionary['store_id'] = sale['pickup_from']['location_id']
    store_dictionary['name'] = sale['pickup_from']['name']
    store_dictionary['street'] = sale['pickup_from']['street']
    store_dictionary['city'] = sale['pickup_from']['city']
    store_dictionary['state'] = sale['pickup_from']['state']
    store_dictionary['zip'] = sale['pickup_from']['zip']
    stores.append(store_dictionary)
    
    # Customers
    customer_dictionary = {}
    customer_dictionary['sale_id'] = sale['sale_id']
    customer_dictionary['customer_id'] = sale['deliver_to']['customer_id']
    customer_dictionary['first_name'] = sale['deliver_to']['first_name']
    customer_dictionary['last_name'] = sale['deliver_to']['last_name']
    customer_dictionary['street'] = sale['deliver_to']['street']
    customer_dictionary['city'] = sale['deliver_to']['city']
    customer_dictionary['state'] = sale['deliver_to']['state']
    customer_dictionary['zip'] = sale['deliver_to']['zip']
    customers.append(customer_dictionary)
    
    # Line Items
    counter = 1
    for line_item in sale['line_items']:
        line_item_dictionary = {}
        line_item_dictionary['sale_id'] = sale['sale_id']
        line_item_dictionary['line_item_id'] = counter
        counter += 1
        for key in line_item:
            line_item_dictionary[key] = line_item[key]
        line_items.append(line_item_dictionary)


## Write CSV

f = open('../Data 2/sales_2.csv', 'w')
dw = csv.DictWriter(f, sales[0].keys())
dw.writeheader()
dw.writerows(sales)
f.close()


f = open('../Data 2/stores_2.csv', 'w')
dw = csv.DictWriter(f, stores[0].keys())
dw.writeheader()
dw.writerows(stores)
f.close()


f = open('../Data 2/customers_2.csv', 'w')
dw = csv.DictWriter(f, customers[0].keys())
dw.writeheader()
dw.writerows(customers)
f.close()


f = open('../Data 2/line_items_2.csv', 'w')
dw = csv.DictWriter(f, line_items[0].keys())
dw.writeheader()
dw.writerows(line_items)
f.close()

## Read CSV

print('Customers')
with open('../Data 2/customers_2.csv') as file:
    customers = list(csv.reader(file))
    for customer in customers[:5]:
        print(customer)

print('\nSales')
with open('../Data 2/sales_2.csv') as file:
    sales = list(csv.reader(file))
    for sale in sales[:5]:
        print(sale)

print('\nStores')
with open('../Data 2/stores_2.csv') as file:
    stores = list(csv.reader(file))
    for store in stores[:5]:
        print(store)

print('\nProducts')
with open('../Data 2/line_items_2.csv') as file:
    products = list(csv.reader(file))
    for line_item in products[:5]:
        print(line_item)
