==> Notebooks 2/6 - Loading.ipynb <==
# **6 - Loading**

## Import Libraries

from sqlalchemy import create_engine, text
import csv
import pandas as pd

## Connect to Database

engine = create_engine('postgresql://bryguan:qwertyuiop@localhost:5432/acme')
connection = engine.connect().execution_options(auto_commit = True)

## Drop Existing Tables

### Customers

query = '''

DROP TABLE IF EXISTS customers_2;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Line Items

query = '''

DROP TABLE IF EXISTS line_items_2;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Product Mapping

query = '''

DROP TABLE IF EXISTS product_mappings_2;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Sales

query = '''

DROP TABLE IF EXISTS sales_2;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Stores

query = '''

DROP TABLE IF EXISTS stores_2;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

## Create Staging Tables

### Customers

with open('/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/customers_2.csv') as file:
    customers = list(csv.reader(file))
    for customer in customers[:5]:
        print(customer)

query = '''

CREATE TABLE customers_2 (
    sale_id numeric(10),
    customer_id numeric(10),
    first_name varchar(25),
    last_name varchar(25),
    street varchar(50),
    city varchar(50),
    state char(2),
    zip_code char(5),
    PRIMARY KEY (sale_id, customer_id)
)

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Line Items 

with open('/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/line_items_2.csv') as file:
    line_items = list(csv.reader(file))
    for line_item in line_items[:5]:
        print(line_item)

query = '''

CREATE TABLE line_items_2 (
    sale_id numeric(10),
    line_item_id numeric(5),
    product_id numeric(10),
    price numeric(5),
    quantity numeric(5),
    taxable boolean,
    PRIMARY KEY (sale_id, line_item_id)
)

'''
try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Product Mappings

with open('/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/product_mappings_2.csv') as file:
    product_mappings = list(csv.reader(file))
    for product_mapping in product_mappings[:5]:
        print(product_mapping)

query = '''

CREATE TABLE product_mappings_2 (
    product_id_1 numeric(5) PRIMARY KEY,
    product_id numeric(10)
)

'''
try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Sales

with open('/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/sales_2.csv') as file:
    sales = list(csv.reader(file))
    for sale in sales[:5]:
        print(sale)

query = '''

CREATE TABLE sales_2 (
    sale_id numeric(10),
    date date,
    subtotal numeric(5),
    tax numeric(5),
    total numeric(5)
)

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Stores

with open('/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/stores_2.csv') as file:
    stores = list(csv.reader(file))
    for store in stores[:5]:
        print(store)

query = '''

CREATE TABLE stores_2 (
    sale_id numeric(10),
    store_id numeric(5),
    name varchar(25),
    street varchar(50),
    city varchar(50),
    state char(2),
    zip_code char(5),
    PRIMARY KEY (sale_id, store_id)
)

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

## Load

### Customers

query = '''

COPY customers_2 (sale_id, customer_id, first_name, last_name, street, city, state, zip_code)
FROM '/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/customers_2.csv'
DELIMITER ','
CSV HEADER;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Line Items

query = '''

COPY line_items_2 (sale_id, line_item_id, product_id, price, quantity, taxable)
FROM '/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/line_items_2.csv'
DELIMITER ','
CSV HEADER;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Product Mappings

query = '''

COPY product_mappings_2 (product_id_1, product_id)
FROM '/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/product_mappings_2.csv'
DELIMITER ','
CSV HEADER;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Sales

query = '''

COPY sales_2 (sale_id, date, subtotal, tax, total)
FROM '/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/sales_2.csv'
DELIMITER ','
CSV HEADER;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

### Stores

query = '''

COPY stores_2 (sale_id, store_id, name, street, city, state, zip_code)
FROM '/Users/bryguan/Documents/GitHub/acme-sales-analysis/Data 2/stores_2.csv'
DELIMITER ','
CSV HEADER;

'''

try:
    with connection.begin():
        connection.execute(text(query))
except Exception as e:
    print(f' Error {e}')

## Verify Load

### Customers

query = '''

SELECT *
FROM customers_2;

'''

display(pd.read_sql_query(query, connection))

### Line Items

query = '''

SELECT *
FROM line_items_2;

'''

display(pd.read_sql_query(query, connection))

### Product Mappings

query = '''

SELECT *
FROM product_mappings_2;

'''

display(pd.read_sql_query(query, connection))

### Sales

query = '''

SELECT *
FROM sales_2;

'''

display(pd.read_sql_query(query, connection))

### Stores

query = '''

SELECT *
FROM stores_2;

'''

display(pd.read_sql_query(query, connection))

## Close Connection

connection.close()
engine.dispose()
