==> Notebooks 2/7 - Validating.ipynb <==
# **7 - Validating**

## Importing Libraries

from sqlalchemy import create_engine
import pandas as pd

## Connect to Database

engine = create_engine('postgresql://bryguan:qwertyuiop@localhost:5432/acme')
connection = engine.connect().execution_options(auto_commit = True)

## Validating Data 2

### Customers 2: Sale ID in Sales 2: Sale ID

query = '''

SELECT
    sale_id
FROM
    customers_2
WHERE
    sale_id NOT IN (
        SELECT
            sale_id
        FROM
            sales_2
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Sales 2: Total = Sales 2: Subtotal + Tax

query = '''

SELECT
    sale_id
FROM
    sales_2
WHERE
    total <> (subtotal + tax);

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Sales 2: Total = Line Items 2: Price * Quantity

query = '''

SELECT
    sale_id
FROM
    sales_2
WHERE
    sales_2.total <> (
        SELECT
            SUM(price * quantity)
        FROM
            line_items_2
        WHERE
            sales_2.sale_id = line_items_2.sale_id
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Stores 2: Sale ID in Sales 2: Sale ID

query = '''

SELECT
    sale_id
FROM
    stores_2
WHERE
    sale_id NOT IN (
        SELECT
            sale_id
        FROM
            sales_2
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Line Items 2: Sale ID in Sales 2: Sale ID

query = '''

SELECT
    sale_id
FROM
    line_items_2
WHERE
    sale_id NOT IN (
        SELECT
            sale_id
        FROM
            sales_2
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Line Items 2: Product ID in Product Mapping 2: Product ID 

query = '''

SELECT
    product_id
FROM
    line_items_2
WHERE
    product_id NOT IN (
        SELECT
            product_id
        FROM
            product_mappings_2
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

## Validating Data 2 IN Data 1

### Customers 2: City, State, ZIP Code in Zip Codes 1: City, State, ZIP Code

query = '''

SELECT
    city,state, zip_code
FROM
    customers_2
WHERE
    (city, state, zip_code) NOT IN (
        SELECT
            city, state, zip_code
        FROM
            zip_codes_1
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Customers 2: First Name, Last Name in Customers 1: First Name, Last Name

query = '''

SELECT
    first_name, last_name
FROM
    customers_2
WHERE
    (first_name, last_name) NOT IN (
        SELECT
            first_name, last_name
        FROM
            customers_1
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Sales 2: Date, Total in Sales 1: Date, Total

query = '''

SELECT
    date, total
FROM
    sales_2
WHERE
    (date, total) NOT IN (
        SELECT
            date, total
        FROM
            sales_1
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Stores 2: City in Stores 1: City

query = '''

SELECT
    city
FROM
    stores_2
WHERE
    city NOT IN (
        SELECT
            city
        FROM
            stores_1
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Product Mappings 2: Product ID 1 IN Product 1: Product ID

query = '''

SELECT
    product_id_1
FROM
    product_mappings_2
WHERE
    product_id_1 NOT IN (
        SELECT
            product_id
        FROM
            products_1
    );

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

## Close Connection

connection.close()
engine.dispose()
