==> Notebooks 2/8 - Querying.ipynb <==
# **8 - Querying**

## Importing Libraries

from sqlalchemy import create_engine
import pandas as pd

## Connect to Database

engine = create_engine('postgresql://bryguan:qwertyuiop@localhost:5432/acme')
connection = engine.connect().execution_options(auto_commit = True)

## Queries

### Gross Sales, Sales Volume, Average Order Value

query = '''

SELECT
    sum(total) as gross_sales,
    count(total) as sales_volume,
    sum(total) / count(*) as average_order_value
FROM 
    sales_2;

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Gross Sales 2, Sales Volume 2 % Gross Sales 1, Sales Volume 1

query = '''

WITH sales_1_data AS (
    SELECT
        SUM(total) AS gross_sales,
        COUNT(total) AS sales_volume
    FROM 
        sales_1
),
sales_2_data AS (
    SELECT
        SUM(total) AS gross_sales,
        COUNT(total) AS sales_volume
    FROM 
        sales_2
)
SELECT
    CAST(sales_2_data.gross_sales AS DECIMAL) / CAST(sales_1_data.gross_sales AS DECIMAL) AS percent_gross_sales,
    CAST(sales_2_data.sales_volume AS DECIMAL) / CAST(sales_1_data.sales_volume AS DECIMAL) AS percent_sales_volume
FROM
    sales_1_data, sales_2_data;

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Customers

query = '''

SELECT
    COUNT(DISTINCT customer_id) as customers
FROM 
    customers_2;

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Customers 2 % Customers 1

query = '''

WITH customers_1_data AS (
    SELECT
        COUNT(DISTINCT customer_id) AS customers
    FROM 
        customers_1
),
customers_2_data AS (
    SELECT
        COUNT(DISTINCT customer_id) AS customers
    FROM 
        customers_2
)
SELECT
    CAST(customers_2_data.customers AS DECIMAL) / CAST(customers_1_data.customers AS DECIMAL) AS percent_customers
FROM
    customers_1_data, customers_2_data;

'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Customers by Sales

query = '''

SELECT
    customer_id,
    first_name,
    last_name,
    count(*) as sales
FROM 
    customers_2
GROUP BY
    customer_id, first_name, last_name
ORDER BY
    sales DESC
    
'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Customers by Quantity

query = '''

SELECT
    customer_id,
    first_name,
    last_name,
    SUM(quantity) as quantity
FROM 
    customers_2
JOIN
    sales_2 ON customers_2.sale_id = sales_2.sale_id
JOIN
    line_items_2 ON sales_2.sale_id = line_items_2.sale_id
GROUP BY
    customer_id, first_name, last_name
ORDER BY
    quantity DESC
    
'''

try:
    with connection.begin():
        display(pd.read_sql_query(query, connection))
except Exception as e:
    print(f' Error {e}')

### Units Sold

query = '''

SELECT SUM(quantity) as units_sold
FROM line_items_2;

'''

connection.rollback()
display(pd.read_sql_query(query, connection))

### Units Sold by Product

query = '''

SELECT
    description as product,
    SUM(quantity) as units_sold
FROM
    line_items_2
JOIN
    product_mappings_2 ON line_items_2.product_id = product_mappings_2.product_id
JOIN
    products_1 ON product_mappings_2.product_id_1 = products_1.product_id
GROUP BY
    description
ORDER BY
    units_sold DESC;

'''

connection.rollback()
display(pd.read_sql_query(query, connection))

## Close Connection

connection.close()
engine.dispose()
